+++ tests/Microsoft.Identity.Test.Unit/PublicApiTests/AuthenticationSchemeTests.cs
@@ -4,13 +4,12 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.AuthScheme;
 using Microsoft.Identity.Client.Cache.Items;
-using Microsoft.Identity.Client.UI;
-using Microsoft.Identity.Test.Common;
+using Microsoft.Identity.Client.OAuth2;
+using Microsoft.Identity.Client.Utils;
 using Microsoft.Identity.Test.Common.Core.Helpers;
 using Microsoft.Identity.Test.Common.Core.Mocks;
 using Microsoft.Identity.Test.Common.Mocks;
@@ -46,7 +45,6 @@ namespace Microsoft.Identity.Test.Unit.PublicApiTests
                 PublicClientApplication app = PublicClientApplicationBuilder.Create(TestConstants.ClientId)
                                                                             .WithHttpManager(httpManager)
                                                                             .BuildConcrete();
-
 
                 app.ServiceBundle.ConfigureMockWebUI();
 
@@ -121,6 +119,26 @@ namespace Microsoft.Identity.Test.Unit.PublicApiTests
             }
         }
 
+        [TestMethod]
+        public async Task MissingAccessTokenTypeInResponse_Throws_Async()
+        {
+            using (MockHttpAndServiceBundle harness = CreateTestHarness())
+            {
+                var requestParams = harness.CreateAuthenticationRequestParameters(TestConstants.AuthorityCommonTenant);
+                var tokenClient = new TokenClient(requestParams);
+
+                harness.HttpManager.AddInstanceDiscoveryMockHandler();
+                var fakeResponse = TestConstants.CreateMsalTokenResponse();
+                fakeResponse.TokenType = null;
+                harness.HttpManager.AddResponseMockHandlerForPost(MockHelpers.CreateSuccessResponseMessage(JsonHelper.SerializeToJson(fakeResponse)));
+                await requestParams.AuthorityManager.RunInstanceDiscoveryAndValidationAsync().ConfigureAwait(false);
+
+                var ex = await Assert.ThrowsExceptionAsync<MsalClientException>(
+                    () => tokenClient.SendTokenRequestAsync(new Dictionary<string, string>())).ConfigureAwait(false);
+                Assert.AreEqual(MsalError.AccessTokenTypeMissing, ex.ErrorCode);
+            }
+        }
+
         private static void ValidateBearerToken(MockHttpManager httpManager, AuthenticationResult result)
         {
             Assert.AreEqual(TestConstants.ATSecret, result.AccessToken);
@@ -160,8 +178,6 @@ namespace Microsoft.Identity.Test.Unit.PublicApiTests
 
             return await builder
                 .ExecuteAsync().ConfigureAwait(false);
-
-
         }
     }
 }
